Name the JupiterOne managed policy
Note the following regarding the use of `ManagedPolicyName`:

> If you specify a name, you cannot perform updates that require replacement of this resource. You can perform updates that require no or some interruption. If you must replace the resource, specify a new name.

We should be okay to specify a name and have customers update the stack resource in the future without specifying a new name. `Description` and `Path` value changes require replacement of the resource and force the requirement of a new name most likely. `PolicyDocument` are all that we would care to update and that will not require a replacement.